code fixes
