+++ apps/admin_app/lib/admin_app/honeydew_workers/export_data_worker.ex
@@ -2,17 +2,17 @@ defmodule Avia.ExportDataWorker do
   @behaviour Honeydew.Worker
 
   alias Snitch.Core.Tools.MultiTenancy.Repo
-  alias AdminAppWeb.Helpers
+  alias AdminAppWeb.Exporter
 
   def export_data(%{"tenant" => tenant, "format" => format, "type" => type, "user" => user}) do
     Repo.set_tenant(tenant)
 
     case format do
       "csv" ->
-        Helpers.csv_exporter(user, type)
+        Exporter.csv_exporter(user, type)
 
       "xlsx" ->
-        Helpers.xlsx_exporter(user, type)
+        Exporter.xlsx_exporter(user, type)
     end
   end
 end
