+++ apps/admin_app/lib/admin_app_web/mailer/data_export_mail.ex
@@ -1,6 +1,6 @@
 defmodule AdminAppWeb.DataExportMail do
   @moduledoc """
-  Composing email to send order/product data export.
+  Composing email to send data export.
   """
 
   import Swoosh.Email
