+++ apps/snitch_core/lib/core/data/model/order.ex
@@ -201,9 +201,15 @@ defmodule Snitch.Data.Model.Order do
   def get_all, do: Repo.all(Order)
 
   @doc """
+  Returns all Orders with the given list of entities preloaded
+  """
+  def get_all_with_preloads(preloads) do
+    Repo.all(Order) |> Repo.preload(preloads)
+  end
+
+  @doc """
     Order related to user.
   """
-
   @spec user_orders(String.t()) :: [Order.t()]
   def user_orders(user_id) do
     query =
